% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.package.list.R
\name{create.package.list}
\alias{create.package.list}
\title{Create package list}
\usage{
create.package.list(docsetroot, overwrite = FALSE)
}
\arguments{
\item{docsetroot}{path to `Documents` dir}

\item{overwrite}{wherether overwrite or not}
}
\description{
Create package list
}

