% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-sqlite-index.R
\name{create.sqlite.index}
\alias{create.sqlite.index}
\title{Create sqlite index for a package}
\usage{
create.sqlite.index(docsetroot, pkg, con)
}
\arguments{
\item{docsetroot}{path to `Documents` dir of docset}

\item{pkg}{package name}

\item{con}{SQLiteConnection}
}
\description{
Create sqlite index for a package
}

