% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docset.init.R
\name{docset.init}
\alias{docset.init}
\title{Create new directory for docset}
\usage{
docset.init(path = getwd())
}
\arguments{
\item{path}{where to create}
}
\description{
Create new directory for docset
}

