% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-docset.R
\name{docset.outdated}
\alias{docset.outdated}
\title{Update docset}
\usage{
docset.outdated(docsetroot, ignored = c())
}
\arguments{
\item{docsetroot}{path to `Documents` dir}
}
\value{
list outdated package,new package
}
\description{
Update docset
}

