% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docset.register.db.R
\name{docset.register.db}
\alias{docset.register.db}
\title{Register SQLite index}
\usage{
docset.register.db(docsetroot, new = TRUE)
}
\arguments{
\item{docsetroot}{path to `Documents` dir of docset}
}
\description{
Register SQLite index
}

