% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.help.html.R
\name{generate.help.from.Rd}
\alias{generate.help.from.Rd}
\title{Generate help html from Rd for each topic}
\usage{
generate.help.from.Rd(docsetroot, pkg)
}
\arguments{
\item{docsetroot}{path to `Documents` dir}

\item{pkg}{package name}
}
\description{
Generate help html from Rd for each topic
}

