% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.help.html.R
\name{generate.help.html}
\alias{generate.help.html}
\title{Generate help html for a package}
\usage{
generate.help.html(docsetroot, pkg)
}
\arguments{
\item{docsetroot}{path to `Documents` dir}

\item{pkg}{package name}
}
\description{
Generate help html for a package
}

