% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.help.html.R
\name{list.of.dir}
\alias{list.of.dir}
\title{List of files of doc dir (used when index.html is not exist)}
\usage{
list.of.dir(docsetroot, pkg, path)
}
\arguments{
\item{docsetroot}{path to `Documents` dir}

\item{pkg}{packagename}

\item{path}{path to pkg/doc/index.html}
}
\description{
List of files of doc dir (used when index.html is not exist)
}

