% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-help-html.R
\name{replace.link.to.description}
\alias{replace.link.to.description}
\title{Replace link}
\usage{
replace.link.to.description(doc)
}
\arguments{
\item{doc}{XML document}
}
\description{
Replace link
}

