% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.help.html.R
\name{replace.logo}
\alias{replace.logo}
\title{Generate help html files for a package}
\usage{
replace.logo(doc)
}
\arguments{
\item{doc}{XML document}
}
\description{
Replace path to R logo, adjust size
}

