% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-help-html.R
\name{update.help.html}
\alias{update.help.html}
\title{Update help html}
\usage{
\method{update}{help.html}(docsetroot, pkg)
}
\arguments{
\item{docsetroot}{path to `Documents` dir}

\item{pkg}{package name}
}
\description{
Update help html
}

